requirements.txt
Django==1.11.15
django-betterforms==1.2
django-crispy-forms==1.7.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-solo==1.1.2
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.1
psycopg2==2.8.2
pytz==2017.2
six==1.11.0
